app/requirements.txt
Flask
Flask-Login
psycopg2-binary
SQLAlchemy
plotly
pandas
numpy